Add: idea file 삭제 및 Report 3,4 제출

+++ src/LEESEUNGRYEOL/report2/Report2_6.java
@@ -5,14 +5,22 @@ public class Report2_6 {
         int fahrenheit = 100;
         Calculate_cel calculate_cel = new Calculate_cel();
         calculate_cel.cal(fahrenheit);
+
+//        int a = 3;
+//        float b = 5/9f;
+//        float c  = a*b;
+//        System.out.println(b);
     }
 }
+
 class Calculate_cel{
     void cal(int fahrenheit)
     {
         float celcius;
-        celcius = (float) 5/9 * (fahrenheit -32);
+        celcius = (5/9f) * (fahrenheit -32);
         int a = (int) (celcius * 1000);
+
+        // b = 올림할껀지 반올림할껀지 판단하는 숫자
         int b = a % 10;
         if (b >= 5)
         {
